wxmaxima already inherited an option to use the pngcairo terminal for gnuplot. This terminal offers an enhanced graphics quality (better fonts and antialiassing) and allows for dashed lines which the simple "png" terminal didn't.
Since old versions of gnuplot didn't support pngcairo the new
terminal was enabled as an option that could be enabled in a
.rc file.

This commit now adds an autodetection that is quite fast and
upon failure falls back to the old png terminal.

In theory it would be possible to cache the result of the
autodetection between wxmaxima runs. But the gain would
be small (the autodetection needs only 50 miliseconds on my
laptop) and would require to write a fallback for the
eventuality that after an update gnuplot no more supports this
terminal.