Added a target "checksums", which generates a checksum file.
I dont see a real value of a checksum file which is on the same server,
but doesnt matter, that was not hard to code.

make package
must be run before. And then:
make checksums

md5 checksums are build in cmake (cmake -E md5sum), for sha1, sha2, sha3, ...
we would need an external tool like openssl.
Do we need checksums for source packages too?