Further clean up the layout code.
I don't really expect that to actually do something. But I still slightly
hope that if the worksheet fits into the window the first time we attempt to
layout it it will actually be placed there.

What this step definitively does is to reduce the number of redraws until
the window is actually usable:

 - The window is drawn
 - The contents of the worksheet and the panes is filled in
 - If a file is to be loaded it is loaded and displayed
 - And then we scroll to the intended position.

Before this change the window was resized and moved after these steps were
completed.