CMake: Updating wxMaxima.pot and *.po in the sourcecode works now.
The procedure should work with the following commands:

cmake .. # in the "build"-directory

will generate all po files and the pot file in the *build* directory.

make update-locale

will *modify* the wxMaxima sourcecode
(Attention: this will break the rule, that an out-of-source-build
should *not* modify anything in the source directory!!!)
so that the POT and PO files are updated (using msgmerge).