CPU efficiency: Moved the auto-hide-brackets and the tooltips code to the idle loop.
This allows us to listen to mouse movements as fast as they happen and then
to actually change the window contents details if there is time between
mouse movement events.

The advantage of doing so is that if deciding which tooltip to show
needs more time than the timespan between mouse movement notifications
we can redraw the screen as fast as we can - and then process all the
mouse movements that have accumulated before we start redrawing the screen
again.