Better Newline handling.
\r\n, \r and \n are all possible line endings. But scanning the whole string
for all of the possibilities all the time is too slow and scanning only
the new text for a \r isn't too good an idea: It might be a packet containing
a part of a string that ends in a \r while the next one will begin in a \n.

The XML Tags don't make use of newlines or are able to handle all
possibilities (at least they should) => only do a search-and-replace on
handling miscellaneous strings.