'install' message files in the directory "${CMAKE_BINARY_DIR}/locale/..." too
(at build time, not at install time), so that running ./wxmaxima-local
(without installing it) can find the message files and i18n works.

It seems to work (much better) with the default language (german in
my case), but if I switch the language in wxMaxima (edit->options)
and restart wxMaxima, the language is not switched.

Strange, because when I look with strace, which language files are found
and opened, they are all read by wxMaxima...