Explicitely kill maxima on closing wxMaxima.
This is the only thing we can do about #1042 from the wxMaxima side.
That means: The only thing we can do without re-inventing a wheel
maxima provides.

If this commit doesn't help please open a ticket at
https://maxima.sourceforge.net explaining the symptoms if this
issue: Perhaps they have an idea on how to canonically start maxima
without having to use a batch file. Or alternatively an idea how to
make maxima automatically close if the system is shut down.