Handle partial writes and errors in SendMaxima
Without this patch, partial writes go unnoticed by wxMaxima,
resulting in Maxima seeing a partial command followed immediately
by whatever the next command to be written happens to be. On
OpenBSD, this seems to happen more often than on other OSes, as I
consistently see this message on startup:

  Maxima encountered a Lisp error:
   Unexpected end of file on #<string-input stream from "(:lis...">.
  Automatically continuing.
  To enable the Lisp debugger set *debugger-hook* to nil.
  Maxima encountered a Lisp error:
   The function READ-WXMAXIMA-VERSION is undefined.
  Automatically continuing.
  To enable the Lisp debugger set *debugger-hook* to nil.

This is caused by wxMaxima's startup commands being incompletely
written.

Regardless of which platforms this bug manifests on, this is the
documented behaviour of the wxSocketBase::Write call:

  https://docs.wxwidgets.org/3.1/classwx_socket_base.html#a7676f1821d2e17db95de2438346355cd
  "Use LastWriteCount() to verify the number of bytes actually
  written."

I am uncertain whether exit(1) is the correct way to handle errors
in this code, but I can't think of a better approach. This method
is called very early after starting Maxima (from SetupVariables),
so trying to automatically recover by restarting Maxima could
easily lead to an infinite loop of Maxima restarts. I think it is
better to fail hard than to risk such a loop.