Match the group type correctly in ChangeCellStyle
Found by a compiler warning:

  /usr/pkg/pobj/wxMaxima-19.08.1/wxmaxima-Version-19.08.1/src/wxMaxima.cpp:8742:10: warning: comparison of two values with different enumeration types in switch statement ('TextStyle' and 'GroupType') [-Wenum-compare-switch]
      case GC_TYPE_HEADING6:
           ^~~~~~~~~~~~~~~~

The intent of this code appears to be to restrict changing the
cell's style to text cells only, but the wrong method call was used
to get the cell's type.

This created a bug whereby changing the cell style appeared to work
correctly for text cells, but non-text cells were broken. To
reproduce:

  1. Go to Cell -> Insert image...
  2. Insert an image cell (with any image)
  3. Click on the text at the top of the image cell to activate it.
  4. Use the drop-down menu to change its type from Maths to any
     other type.
  5. The image cell will be replaced with a non-image cell, and the
     image will vanish from view.