Got rid of another gcc warning
When optimizing the code the definition of maxima's built in commands
was so long that GCC gave up on optimizing it along with the rest
of the function it was done in. I've therefore split it into a separate
function and (because the file was still way too long) moved this
function into a separate file.