The out-of-tree-build warning interferred with bash autocompletion
bash provides autocompletion for the command

    cmake <path to the project>

...but every time bash tries to initiate it the out-of-tree-build
warning popped up making the result hard to read.

I guess the warning is no more strictly necessary any more, because
compiling now works fine in both configurations. ...but if the
warning could be output at the beginning of the build instead of
every time cmake is run I would be happy with that, too.