wxMaxima now uses the new wxMathML.lisp
Since ArchLinux, Debian SuSE, Mac with Fink, Mac with MacPorts, Windows using
Cygwin, Windows using the installer, Mac using the MacBundle, portable apps
etc. each use their own direcectory layout and these directory layouts tend
to change from time to time wxMaxima (if it can help it) doesn't search for
files but bundles a gnuzip'ed version within the .exe file.

For making wxMaxima bundle the new data/wxmathml.lisp
data/generate_wxmathml.sh needs to be executed. Perhaps I should describe that
in a README.md file, as a 2nd thought...

This commit is a small part of the solution for #1195. The rest needs to be
done in src/MathParser.cpp - but with modifying the C++ code to speak the
same language as the lisp part I feel confident to be able to help.