Explicitely initialize and deinit wxSocketBase
In theory that shouldn't be necessary but the MSW/Mac problems
with connecting maxima look like a race condition so it might be wise
to disrupt the timing. Initializing and deinitializing sockets with
every window is allowed, according to the documentation: Only the
first init and the last uninit are effective.