In MathML output for index properties, honor the display_index_separator property.
The display_index_separator can be a string, including an empty string,
or false, indicating the default separator, or any expression. If not a
string and not false, the property value is coerced to a string. (I
don't know how useful that will be, but I just wanted to make sure the
output is reasonable in that case.)

Examples:

    declare_index_properties (A, [postsuperscript, postsuperscript, presubscript, presubscript]);
    A[w, x, y, z];
    put (A, ";", display_index_separator);
    A[w, x, y, z];
    put (A, " ", display_index_separator);
    A[w, x, y, z];
    put (A, "", display_index_separator);
    A[w, x, y, z];
    put (A, false, display_index_separator);
    A[w, x, y, z];
    put (A, 'foo, display_index_separator);
    A[w, x, y, z];