Only add an anchor to the Maxima manual, if a keyword was given.
That is another idea for bug #1221. Maybe the browser searches the
file name with the anchor as part of the filename.

For the general manual it is problably a good idea, since one
usually starts reading at the beginning of the manual (where the
table of contents is).
But if that is the 'solution', that would break the context sensitive
help...