Another bug found by MSoegtropIMC: AdjustSize didn't mark itself as done
AdjustSize is not immediately called when the worksheet size changes.
Instead it is called by the idle task when nothing else is to do and
before the worksheet is redrawn. Unfortunately AdjustSize didn't mark
itself as "done" and therefore was re-done on the next redraw of
the worksheet - or the next time wxMaxima went idle. Which caused
heavy CPU load.

No idea why wxMaxima still worked fine on Linux and MSW, though...
...but this should be another step for #1220.