(Hopefully) fixed updating the manual translations with po4a-updatepo.
make update-locale-manual
will update the language files in the source tree.

Currently the option "neverwrap" is not used, it should be, so that
for example captions are not splitted in a part which is a caption and
a part, which is not. But that causes problems with line endings.

Maybe a better solution is to just extend the default length. I will
do more experiments for that problem.