Change "Export as pdfLaTeX" to "Export as LaTeX".
The exported document can be compiled not only with
pdflatex but also with xelatex, lualatex and latex.
And currently lualatex works better than pdflatex,
so no need to explicitly mention pdflatex.

And I assume other programs, which offer a LaTeX
export function also call it "export to LaTeX" (or maybe
"export to TeX"), not "export to pdfLaTeX".