AltGr is kind of Shift+Alt, but isn't
Alt+a is an ordinary "a" that we want to use as a hotkey, not
as an "a". And I hope no-one has a keyboard that requires
the ALT key to be pressed for typing latin letters as I somehow
seem to run out of ideas on how to decide which keypress is meant to
produce letters.

At least #1273 should be resolved now.