Don't use globbing for finding the source files
Resolves #1316

No idea if we should do the same for the translation languages:
The cmake manual was strictly against globbing as long as
adding or removing source files wasn't recognized by cmake.

With source files one cannot forget to add files to the sources
list without breaking the compilation. Well, one actually can
if a source file isn't used on every platform. But with translations
without globbing it is easy to forget to add a new translation.
Not using globbing here might therefore be a potential source of
errors the CIs cannot possibly see.