Work around when UpdateYPosition ends up called without a valid position chain.
This is just to prevent the assertion and is not a proper fix.

The call chain was GroupCell::AppendOutput()-> Recalculate()-> RecalulateHeight()-> UpdateYPositionList()-> UpdateYPosition(). This happened when an `ImgCell` was being added to the output, from within the xml parser, and none of the previous cells had had their position calculated yet.

Most likely `AppendOutput()` should not do a `Recalculate()` and only schedule it for later. But then the results of line breaking in the output cells don't have their Y positions updated properly. So there probably is an `UpdateYPoisitionList` call that's missing after the line breaking process.