Include first the directory info, then locales.
In info/CMakeLists.txt pandoc is searched, but it is
used in locales/CMakeLists.txt too.

make update-locale-manual-in-source
(which is re-enabled with this patch)
still does not work correctly on my PC, instead of

type: Fenced code block (maxima)
it declares code blocks as
type: Plain text

Therefore I can not update the po files.

Maybe a bug in my software toolchain, it seems to work
correctly on other computers.