Check for a suitable appstream-util version for the check_appstream_file test.
I assume, the issue with our 'invalid' ID (io.github.wxmaxima_developers.wxMaxima)
with older versions was fixed in version 0.7.2: "Do not check the suffix of <id> tags"
(https://github.com/hughsie/appstream-glib/blob/appstream_glib_0_7_2/NEWS)
The ID should be unique, so we should probably not change it (it often seems
to have the suffix '.desktop'), so check for a newer appstream-util, which
does not count this as error.

Re-enable the appdata- and desktop-filechecks on Ubuntu 16.04 by installing
the required programs (which were disabled with commit b801cb7).
(one should work and on one we should see, that an older version of
appstream-util was detected.