Change the name for wxwidgets translations from wxMaxima-wxstd to wxstd.
The wxwidgets locale files were called "wxMaxima-wxstd" in the Windows
installer, but that will be renamed to the standard name for wxWidgets
locale files. And with that name the locale files for wxWidgets are
also available using on operating systems (where wxWidgets may be
installed using the package management).
[I found no example, where a translation from the wxWidgets locale
file was used, but it does not hurt...]