PDF manual: Change fonts to DejaVu (better unicode support).
As mentioned in Github discussions
(https://github.com/wxMaxima-developers/wxmaxima/discussions/1539)
the generation of PDF manuals in chinese and ukrainian did not work,
because their characters were not included in the default font.
(and greek letters were missing too, which was a problem for all
languages).
The DejaVu font family includes these characters, which solves
that problem.