Update COPYING with information about included components.
The file COPYING is now shown in wxMaxima's license dialog.
Links to details about the licenses of wxMaxima and used
components are included (and one can open them with a
web browser (commit [3070c1a5]).

No one reads the full GPL license text and the FSF
recommends to show the short version and a pointer (link)
to the full text too...