Fix some assertions with newer wxWidgets (reported as bug #1658).
wxWidgets added some assertions for useless flags, e.g.:

assert "CheckSizerFlags(!((flags) & (wxALIGN_RIGHT)))" failed in DoInsert():
wxALIGN_RIGHT will be ignored in this sizer: only vertical alignment flags
can be used in horizontal sizers.

This is not so catastrophic as written in the bug report, the user can click
on "continue" and wxMaxima works - it does not crash.

Found the same problem in Equations->Boundary value problem.

There might be other problems of the same type too, but I tried
most menus and did not find another.