Resolved the recursive wxYield which broke the Ubuntu ci builds
If we need to wait for the background process that indexes the Maxima
manual before we can process the next event we do two things:

 * We block all events the user can cause
 * and we call wxYield on a regular basis so the GUI can process
   redraw events and tell the operating system that it hasn't stopped
   responding.

But the wxYield seems to be able to trigger processing of network
events which can trigger events that require us to wait for the
manual being indexed - which in turn causes another wxYield that,
because it is inside a wxYield, is disallowed.

Not calling a wxYield if we already wait for the background task
when we start checking if we need to do so resolves that problem.