Ensure unique events for more events
The proper (but ugly) solution to #1726: wxMaxima uses element IDs
for identifying buttons, menu items and tries to assign them an
unique ID by assigning them IDs that are higher than any one
wxWidgets uses internally. But every element that is dynamically
generated with the ID -1 or wxID_ANY is automatically assigned a
new ID by wxWidgets - which doesn't know which IDs we have started
using in the meantime => sooner or later we got a ID collision.

The solution is to let wxWidgets generate the IDs we care about
dynamically, too, so it knows that these IDs are already assigned
to some item. But that solution is ugly as as soon as the IDs get
dynamic C++ doesn't allow us to use switch() statements, any more.

Resolves #1726