Re-enable the generation of Doxygen documentation using github actions.
The documentation on http://www.physikbuch.de/wxMaxima/html/index.xhtml
(linked on our websites) does no longer exist. And it might be better
to have it on our primary website.

Denvercoder1 has implemented my feature request, that deployment
should be done to a subdirectory only, not overwriting the complete
website. Let's see, if it works.