Fix the shape of the integral sign. (for wxWidgets >= 3.1.6)
FIXME: Now the shape is perfect (it's a SVG graphic), but the
position/size is wrong!!! Very wrong!
Will try to fix that later, but why is point, m_signWidth and
m_signHeight not the correct size/position?

This is a first part for issue #1812. (and only for recent
wxWidgets versions).