Fix a bung in "Copy for Matlab/Octave" for Text cells.
First
wxString text = ToString();
is called - which already add quotes at the beginning and the end.

If we add additional quotes, then the output of e.g. the command
"load(ctensor);" would be copied (for Matlab/Octave) with double
quotes, eg.
""/usr/share/maxima/5.43.2/share/tensor/ctensor.mac""
which is of course not a valid Octave string.

Hope, that change does not have side effects..