wxbuild_info(): Format Build Date and System type with ~30a.
Consistent with the other description outputs, instead of
a special number of spaces. So, if one needs more space for
the description of the items, it should be easier to increase
the 30 in ~30a.

But that nice formatting only works, if a monospaced font is
used, which is not neccessary the case.