Remove the <annotation ...> from the Copy as MathML output.
We do not need it (there is no import from MathML in wXmaxima)
and it seems, that that element (although it is valid XML)
hinders Libreoffice (6.4, Ubuntu 20.04) from importing the
XML as math formula. It works without the annotation element.

Other <annotation>-elements often seen, e.g.
  <annotation encoding="application/x-maple">
  <annotation encoding="application/x-tex">
cause simliar troubles, so it is not just the "wxmaxima"
annotation.