Fix crash when exporting worksheets with animations to TeX.
We got just a segfault, when a worksheet with animations was
exported to TeX. Now we just output:
"[animated graphics - not shown in TeX export]"

Yes, there are packages for embedding multimedia (mp4, flash video) in TeX
documents (e.g. media9) but they only work with the propietary Adobe Reader.

And TeX docments are rather generated for printing on paper, if one wants
a multimedia capable output, one should use HTML export (which does include
the animation (as animated gif)).