History: do not 'wrap over'.
Use the same behaviour as readline enabled programs - if
the first entry in the history list is reached stay at the
first entry.
Also move to the first entry in the history list, if it is 'selected'.
(current >= 0 instead of current >0).

There is still an issue with ALT-CursorDown - the selected command
is printed on a new line and I do not see, why...