Use (standard) wxCMD_LINE_OPTION_HELP for the help message.
Stdout is filtered (on Windows), so the idea, that if one
enters a command line option on the command line, then the
output should be on the command line, does not work.

Show the help message (on Windows) in the usual message box
(as other programs). If an unknown option was entered, the
usage was already shown in a message box.