Output the '--version' using wxMessageOutput.
Before it was *not* shown on Windows; now it should
be printed on the console, if it is available, otherwise
in a message box.

The output (of --help and --version) is now on stderr, but
that seems not so uncommon, e.g. "perl --version", "python --version",
"java -version", "bzip2 --version", ... - many commands
output their version number on stderr.