Revert "Revert "Remove unused code. po4a takes care of the update of the po files now.""
Yes, a double revert. This code *is* not necessary with the switch
to po4a, which does the update of the manual po/pot files.

I just thought, because of an other error (introduced in commit
ea02af28a0d597373192cfd2ee9effc513fbe482, corrected in commit
1ac1e571b32688c8a247beeb82f30194612c1f84 that that code removal
does not work.

This reverts commit 68a23ad00021d5fcb74159ec6a729b84db5c5d46.