wxMaxima was not responsive in the 1st minute.
One could enter a command (e.g. 1+1) and send it
to Maxima, but the answer was given approx.
1 minute later. Afterwards everything worked okay.

That was really disturbing me - and no I tried
a bisect.
git bisect showed, that the problematic commit
was 7daf5f6f5d0c6a41fc25bdbb5e04d9e189ac62f2.

I am not sure, what was the problem (and why it
worked after approx 1 minute), but that change
to wxMathML.lisp seems to solve the issue.