Github workflow: remove test once, rename jobs.
There is no specific 'show help'-test, that can
fail, if webview is not included, so skip
the tests for this run.
They are done for the second run.

And rename the first job to "compile_latest_and_test",
so that is clear, that we do complilation and testing
in one job.