Change "Allow to access" without a specific manual.
We do not store two different permissions for accessing
the online manual for wxMaxima and Maxima. Just one
permission, if we might access online (wxMaxima or Maxima)
manuals.

So change "Allow to access a online manual for maxima?"
to the general question "Allow to access an online manual?".

That was mentioned in issue: https://github.com/wxMaxima-developers/wxmaxima/issues/1981