More work on killing Maxima processes.
wxSIGTERM does not work on Windows, one must use wxSIGKILL:
ttps://github.com/wxWidgets/wxWidgets/issues/15356

But it seems that does not work correctly too - either it
does not send the right signal to child processes or it
does not kill the children of the childs.

So, use taskkill /PID nn /T /F now.