Use the (standard) wxWidget log window for logging.
Less code to maintain, more features (e.g. clear log, save log).
In "Release" builds the log window will be hidden (and can be enabled
using View->Show log window, otherwise it is shown.

FIXME: If the window was closed (using the close icon in the upper
right corner), the checkbox in the menu is not unchecked, it needs
two clicks on View->Show log window to re-enalbe it.

This improves the solution for issue
https://github.com/wxMaxima-developers/wxmaxima/issues/1954