Kill the log window when closing wxMaxima. (Fixes: #1999).
Note: If you open multiple documents and close one wxMaxima session,
the log window is destroyed and the toggle log window menu does no longer work.
But I think, the log window is mainly for developers, so that is not a severe
issue for end users. And if someone uses '--logtostderr', log messages are
still printed there.