Remove setting lmxchar/rmxchar in wxMathML.lisp. Fixes: #1926.
Setting lmxchar/rmxchar (used for displaying a matrix in text
mode) did not work (with setq) as shown in issue #1926.
It would work with:
(defmvar $lmxchar "("  "Character used for drawing the left edge of a matrix.")
(defmvar $rmxchar ")"  "Character used for drawing the right edge of a matrix.")

But why modify it at all? Maxima provides a reasonable default ("["/"]").
And people probably expect the same, if they are using command line
Maxima or wxMaxima.
If someone wants another char, one can even set another default value
in maxima-init.mac.
No need to modify it - in my opinion.