Close the log window after closing the last wxMaxima window.
Now the log window is not destroyed, before the last
wxMaxima window. Introduced a windowcounter for that.

We do no longer need to 'Reparent' the window, so the
log window will not be on top of the wxMaxima window on
Windows.
(wxWidgets issue https://github.com/wxWidgets/wxWidgets/issues/25258)

TODO: C++ify the code.