Fix wrong char encodings.
"\xDCB6" should probably be "\uDCB6" (unicode, not
hex DC (Latin capital letter U with diaeresis), than "B", than "6")),
but that is a Surrogate Code Point, not a valid utf8 char - and has
nothing to do with non-breakable-spaces.

The unicode sign is "\u2212", not "\x2212" ("12).

This fixes the core dump reported in issue
https://github.com/wxMaxima-developers/wxmaxima/issues/2012.

We still have an issue, when wxWidgets is configured with
--enable-utf8=yes (but not: --enable-utf8only=yes).

But it works now for both flags (and of course without any of
these flags).