Fix a crash when copying a input cell (collapsed to only show the input).
Got an
assert "(argtype & (wxFormatStringSpecifier<T>::value)) == argtype" failed in wxArgNormalizer(): format specifier doesn't match argument type

The result of GetTextStyle() is int8_t, which is not a long integer.
Just an integer works.
The next argument to wxString::Format is most likely useless and can be
removed, but I want to fix the found bug with a small patch.

Found this bug while trying to debug issue #2023.