Correct socket flags for (not yet released) wxWidgets 3.3.2.
In issue https://github.com/wxMaxima-developers/wxmaxima/issues/2028
it was reported, that the communication Maxima/wxMaxima does not work on
OpenBSD. (verified by me).
The suggested solution, does not work with the most recent wxWidgets (3.3.1),
but the current development version has included a fix:
https://github.com/wxWidgets/wxWidgets/commit/943e4fb18d7b2a3b717628dcc576132837945905
which will be included in the next release.

With a current self compiled wxWidgets (development) version, and that fix
the communication works.
I did only the first change
SetFlags(wxSOCKET_NOWAIT_READ | wxSOCKET_WAITALL_WRITE)
the second (we probably want read and write operations to be finished, before
we terminate the connection) and third seem not necessary (in my opinion...).