Remember the "log window" state (fixes #2033).
Now if the user toggled the log window off, it is not
shown by default at the next start. If it was toggled on
(shown), it is shown by default at the next start.

I think, that is the solution, that issue #2033 requested.

(It seems sometimes one must click twice to "Toggle log window"
to show it again at start, not sure, whats the reason...
But if it does not work, one will probably try it a second time -
and then it works. That problem seems not occur every time here.)